+++ src/spaczz/exceptions.py
@@ -21,6 +21,10 @@ class PatternTypeWarning(Warning):
     """It warns if the spaczz pattern does not have a valid pattern type."""
 
 
+class PipeDeprecation(Warning):
+    """Warns that `matcher.pipe` methods are now deprecated."""
+
+
 class MissingVectorsWarning(Warning):
     """It warns if the spaCy Vocab does not have word vectors."""