+++ src/spaczz/search/tokensearcher.py
@@ -7,7 +7,8 @@ import regex
 from spacy.tokens import Doc, Token
 from spacy.vocab import Vocab
 
-from ..process import FuzzyFuncs, n_wise
+from ..process import FuzzyFuncs
+from ..util import n_wise
 
 
 class TokenSearcher: