+++ tests/test_matcher/test_tokenmatcher.py
@@ -1,6 +1,8 @@
 """Tests for tokenmatcher module."""
 from __future__ import annotations
 
+import pickle
+
 import pytest
 from spacy.language import Language
 from spacy.tokens import Doc, Span
@@ -12,34 +14,34 @@ def add_name_ent(
     matcher: TokenMatcher, doc: Doc, i: int, matches: list[tuple[str, int, int, None]]
 ) -> None:
     """Callback on match function. Adds "NAME" entities to doc."""
-    _match_id, start, end, _ = matches[i]
+    _match_id, start, end, _details = matches[i]
     entity = Span(doc, start, end, label="NAME")
     doc.ents += (entity,)
 
 
 @pytest.fixture
-def matcher(model: Language) -> TokenMatcher:
+def matcher(nlp: Language) -> TokenMatcher:
     """It returns a token matcher."""
-    matcher = TokenMatcher(vocab=model.vocab)
+    matcher = TokenMatcher(vocab=nlp.vocab)
     matcher.add(
         "DATA",
         [
             [
                 {"TEXT": "SQL"},
                 {"LOWER": {"FREGEX": "(database){s<=1}"}},
-                {"LOWER": {"FUZZY": "access"}, "POS": "NOUN"},
+                {"LOWER": {"FUZZY": "access"}},
             ],
             [{"TEXT": {"FUZZY": "Sequel"}}, {"LOWER": "db"}],
         ],
     )
-    matcher.add("NAME", [[{"TEXT": {"FUZZY": "Garfield"}}]])
+    matcher.add("NAME", [[{"TEXT": {"FUZZY": "Garfield"}}]], on_match=add_name_ent)
     return matcher
 
 
 @pytest.fixture
-def example(model: Language) -> Doc:
+def doc(nlp: Language) -> Doc:
     """Example doc for search."""
-    return model(
+    return nlp(
         """The manager gave me SQL databesE acess so now I can acces the Sequal DB.
         My manager's name is Grfield.
         """
@@ -54,7 +56,7 @@ def test_adding_patterns(matcher: TokenMatcher) -> None:
             "pattern": [
                 {"TEXT": "SQL"},
                 {"LOWER": {"FREGEX": "(database){s<=1}"}},
-                {"LOWER": {"FUZZY": "access"}, "POS": "NOUN"},
+                {"LOWER": {"FUZZY": "access"}},
             ],
             "type": "token",
         },
@@ -71,7 +73,7 @@ def test_adding_patterns(matcher: TokenMatcher) -> None:
     ]
 
 
-def test_add_without_sequence_of_patterns_raises_error(matcher: TokenMatcher,) -> None:
+def test_add_without_list_of_patterns_raises_error(matcher: TokenMatcher) -> None:
     """Trying to add non-sequences of patterns raises a TypeError."""
     with pytest.raises(TypeError):
         matcher.add("TEST", [{"TEXT": "error"}])  # type: ignore
@@ -83,12 +85,12 @@ def test_add_with_zero_len_pattern(matcher: TokenMatcher) -> None:
         matcher.add("TEST", [[]])
 
 
-def test_len_returns_count_of_labels_in_matcher(matcher: TokenMatcher,) -> None:
+def test_len_returns_count_of_labels_in_matcher(matcher: TokenMatcher) -> None:
     """It returns the sum of unique labels in the matcher."""
     assert len(matcher) == 2
 
 
-def test_in_returns_bool_of_label_in_matcher(matcher: TokenMatcher,) -> None:
+def test_in_returns_bool_of_label_in_matcher(matcher: TokenMatcher) -> None:
     """It returns whether a label is in the matcher."""
     assert "DATA" in matcher
 
@@ -98,9 +100,9 @@ def test_labels_returns_label_names(matcher: TokenMatcher) -> None:
     assert all(label in matcher.labels for label in ("DATA", "NAME"))
 
 
-def test_vocab_prop_returns_vocab(matcher: TokenMatcher, model: Language) -> None:
+def test_vocab_prop_returns_vocab(matcher: TokenMatcher, nlp: Language) -> None:
     """It returns the vocab it was initialized with."""
-    assert matcher.vocab == model.vocab
+    assert matcher.vocab == nlp.vocab
 
 
 def test_remove_label(matcher: TokenMatcher) -> None:
@@ -119,9 +121,9 @@ def test_remove_label_raises_error_if_label_not_in_matcher(
         matcher.remove("TEST")
 
 
-def test_matcher_returns_matches(matcher: TokenMatcher, example: Doc) -> None:
+def test_matcher_returns_matches(matcher: TokenMatcher, doc: Doc) -> None:
     """Calling the matcher on a `Doc` object returns matches."""
-    assert matcher(example) == [
+    assert matcher(doc) == [
         ("DATA", 4, 7, None),
         ("DATA", 13, 15, None),
         ("NAME", 22, 23, None),
@@ -145,10 +147,10 @@ def test_matcher_warns_if_unknown_pattern_elements(nlp: Language) -> None:
         matcher(doc)
 
 
-def test_matcher_uses_on_match_callback(matcher: TokenMatcher, example: Doc) -> None:
+def test_matcher_uses_on_match_callback(matcher: TokenMatcher, doc: Doc) -> None:
     """It utilizes callback on match functions passed when called on a Doc object."""
-    matcher(example)
-    ent_text = [ent.text for ent in example.ents]
+    matcher(doc)
+    ent_text = [ent.text for ent in doc.ents]
     assert "Grfield" in ent_text
 
 
@@ -182,7 +184,7 @@ def test_matcher_pipe_with_matches(nlp: Language) -> None:
     )
     matcher = TokenMatcher(nlp.vocab)
     matcher.add(
-        "DRAGON", [[{"TEXT": {"FUZZY": "Korvold"}}], [{"TEXT": {"FUZZY": "Prossh"}}]]
+        "DRAGON", [[{"TEXT": {"FUZZY": "Korvold"}}], [{"TEXT": {"FUZZY": "Prossh"}}]],
     )
     output = matcher.pipe(doc_stream, return_matches=True)
     matches = [entry[1] for entry in output]
@@ -197,7 +199,7 @@ def test_matcher_pipe_with_matches_and_context(nlp: Language) -> None:
     )
     matcher = TokenMatcher(nlp.vocab)
     matcher.add(
-        "DRAGON", [[{"TEXT": {"FUZZY": "Korvold"}}], [{"TEXT": {"FUZZY": "Prossh"}}]]
+        "DRAGON", [[{"TEXT": {"FUZZY": "Korvold"}}], [{"TEXT": {"FUZZY": "Prossh"}}]],
     )
     output = matcher.pipe(doc_stream, return_matches=True, as_tuples=True)
     matches = [(entry[0][1], entry[1]) for entry in output]
@@ -205,3 +207,21 @@ def test_matcher_pipe_with_matches_and_context(nlp: Language) -> None:
         ([("DRAGON", 4, 5, None)], "Jund"),
         ([("DRAGON", 4, 5, None)], "Jund"),
     ]
+
+
+def test_pickling_matcher(nlp: Language) -> None:
+    """It pickles the matcher object."""
+    matcher = TokenMatcher(nlp.vocab)
+    matcher.add("NAME", [[{"TEXT": {"FUZZY": "Ridley"}}, {"TEXT": {"FUZZY": "Scott"}}]])
+    bytestring = pickle.dumps(matcher)
+    assert type(bytestring) == bytes
+
+
+def test_unpickling_matcher(nlp: Language) -> None:
+    """It unpickles the matcher object."""
+    matcher = TokenMatcher(nlp.vocab)
+    matcher.add("NAME", [[{"TEXT": {"FUZZY": "Ridley"}}, {"TEXT": {"FUZZY": "Scott"}}]])
+    bytestring = pickle.dumps(matcher)
+    matcher = pickle.loads(bytestring)
+    doc = nlp("Rdley Scot was the director of Alien.")
+    assert matcher(doc) == [("NAME", 0, 2, None)]